Specify correct version number for get-user-teams-memberships
I'd previously updated the `tspascoal/get-user-teams-membership` dependency in
[1] to fix Node 16 warnings on Github Action workflow runs.

Despite following the repo's README instructions to use the "v3" ref, I'd
noticed that we were still getting Node 16 warnings.

It turns out that the v3 release tag [2] doesn't include in the Node 16 fix, but
the v3.0.0 release tag [3] does, so this commit updates to use that ref
instead.

[1] 9fab6aa9f470a42f49771597373edd6bcf1d3003
[2] https://github.com/tspascoal/get-user-teams-membership/releases/tag/v3
[3] https://github.com/tspascoal/get-user-teams-membership/releases/tag/v3.0.0